Move sstOverlay from stores to ui and some sst related properties from ui to sstOverlay.
